+++ src/Storage/Storage.Management/ChangeLog.md
@@ -25,7 +25,14 @@
     -  `Start-AzStorageBlobCopy`
     -  `Get-AzDataLakeGen2Item`
 * Fixed [#12592]: Fix download blob fail when related sub directory not exist.
-	- `Get-AzStorageBlobContent`
+    -  `Get-AzStorageBlobContent`
+* Support Set/Get/Remove Object Replication Policy on Storage accounts
+    - `New-AzStorageObjectReplicationPolicyRule`
+    - `Set-AzStorageObjectReplicationPolicy`
+    - `Get-AzStorageObjectReplicationPolicy`
+    - `Remove-AzStorageObjectReplicationPolicy`
+* Support enable/disable Changefeed on Blob Service of a Storage account
+    - `Update-AzStorageBlobServiceProperty`
 
 ## Version 2.4.0
 * Supported create container/blob Sas token with new permission x,t